rename QuizLayout to QuizPage
