+++ resources/js/Pages/Admin/QuizDemo.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import QuizLayout from '@/Layouts/QuizLayout.vue'
+import QuizPage from '@/components/UserQuiz/QuizPage.vue'
 import { converter } from '@/Helpers/Converter'
 
 const props = defineProps<{ quiz: Quiz }>()
@@ -7,7 +7,7 @@ const props = defineProps<{ quiz: Quiz }>()
 
 <template>
   <div class="w-full bg-primary text-white text-center font-bold mb-4 p-1">Wersja demonstracyjna</div>
-  <QuizLayout
+  <QuizPage
     :user-quiz="converter.quiz.adminToUser(props.quiz)"
     :request-close-quiz="{ method: 'get', href: '/admin/quizzes/' }"
   />
