+++ resources/js/Pages/User/Quiz.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import QuizLayout from '@/Layouts/QuizLayout.vue'
+import QuizPage from '@/components/UserQuiz/QuizPage.vue'
 import axios from 'axios'
 import { ref } from 'vue'
 import Button from '@/components/Common/Button.vue'
@@ -17,12 +17,12 @@ function handleAnswer(question: UserQuestion, selectedAnswer: number) {
 </script>
 
 <template>
-  <QuizLayout 
+  <QuizPage
     :user-quiz="userQuiz"
     :request-close-quiz="{ method: 'post', href: `/quizzes/${userQuiz.id}/close` }"
     @answer="handleAnswer"
   />
-  
+
   <MessageBox :open="networkErrorMessage" @close="networkErrorMessage = false">
     <template #title>Nie udało się wysłać odpowiedzi</template>
